requirements.txt
click==8.1.3
Flask==2.0.0
mysql_connector_repackaged==0.3.1
python-dotenv==0.21.0
setuptools==63.1.0
WTForms==2.3.1
